developed pagination cart

+++ models/cart.py
@@ -5,7 +5,7 @@
 # note that the item is NOT reserved or blocked so that the availability of the item
 # needs to be checked again during checkout
 from pydantic import BaseModel
-from sqlalchemy import Integer, ForeignKey
+from sqlalchemy import Integer, ForeignKey, BigInteger
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from models.base import Base
@@ -15,7 +15,7 @@ class Cart(Base):
     __tablename__ = "carts"
 
     id: Mapped[int] = mapped_column(Integer, primary_key=True)
-    user_id: Mapped[int] = mapped_column(Integer, ForeignKey('users.id'), nullable=False)
+    user_id: Mapped[int] = mapped_column(BigInteger, ForeignKey('users.id'), nullable=False)
     # ForeignKey('users.id') означает, что мы можем вставить в carts.user_id только существующие значения из users.id
     # А также ForeignKey указывает кто является родителем, а кто ребёнком
     cart_items: Mapped[list["CartItem"]] = relationship("CartItem", back_populates="cart", cascade="all, delete-orphan")
